Creating a TaskFragment and moving all logics from TaskActivity into it
