+++ app/src/main/res/layout/top_action_bar.xml
@@ -1,12 +1,11 @@
 <?xml version="1.0" encoding="utf-8"?>
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools">
-    
+    xmlns:app="http://schemas.android.com/apk/res-auto">
+
     <data>
         <variable
-            name="activity"
-            type="com.example.todolistproject.view.TaskActivity" />
+            name="fragment"
+            type="com.example.todolistproject.view.fragment.TaskFragment" />
     </data>
     
     <androidx.constraintlayout.widget.ConstraintLayout
@@ -46,7 +45,7 @@
                 android:layout_marginTop="15dp"
                 android:layout_marginEnd="15dp"
                 android:layout_marginBottom="15dp"
-                android:onClick="@{() -> activity.openAddTaskDialog()}"
+                android:onClick="@{() -> fragment.openAddTaskDialog()}"
                 app:layout_constraintBottom_toBottomOf="parent"
                 app:layout_constraintEnd_toEndOf="parent"
                 app:layout_constraintTop_toTopOf="parent"
